fix: fix Linux package scripts (#800)
* reload systemd daemon to make it aware of an updated service file
  after package update
* restart service to use the new binary after a package update
* don't stop and disable service after rpm package update

fixes #799

+++ distributions/otelcol-otlp/preremove.sh
@@ -14,7 +14,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-if command -v systemctl >/dev/null 2>&1; then
-    systemctl stop otelcol-otlp.service
-    systemctl disable otelcol-otlp.service
+if [ "$1" != "1" ]; then
+    if command -v systemctl >/dev/null 2>&1; then
+        systemctl stop otelcol-otlp.service
+        systemctl disable otelcol-otlp.service
+    fi
 fi
